Hoist url params resolution to entrypoints
