--- src/Playroom/Preview.css.ts
@@ -1,11 +1,0 @@
-import { sprinkles } from './sprinkles.css';
-
-export const renderContainer = sprinkles({
-  position: 'relative',
-  zIndex: 0,
-});
-
-export const splashScreenContainer = sprinkles({
-  position: 'relative',
-  zIndex: 1,
-});
