+++ src/Playroom/Preview/SplashScreen/SplashScreen.tsx
@@ -1,7 +1,7 @@
 import classnames from 'classnames';
 import { useState, useEffect } from 'react';
 
-import { Logo } from '../Logo/Logo';
+import { Logo } from '../../Logo/Logo';
 
 import * as stylesheet from './SplashScreen.css';
