+++ src/frame.tsx
@@ -1,10 +1,11 @@
-import Frame from './Playroom/Frame';
-import { SendErrorMessage } from './Playroom/Frames/frameMessaging';
+import Frame from './Playroom/Frame/Frame';
+import { SendErrorMessage } from './Playroom/Frame/frameMessaging';
 import playroomComponents from './components';
 import PlayroomFrameComponent from './frameComponent';
 import { renderElement } from './render';
 import playroomThemes from './themes';
 import { hmrAccept } from './utils/hmr';
+import { UrlParams } from './utils/params';
 
 const outlet = document.createElement('div');
 document.body.appendChild(outlet);
@@ -15,12 +16,18 @@ const renderFrame = ({
   FrameComponent = PlayroomFrameComponent,
 } = {}) => {
   renderElement(
-    <Frame
-      components={components}
-      themes={themes}
-      FrameComponent={FrameComponent}
-      ErrorComponent={SendErrorMessage}
-    />,
+    <UrlParams themes={themes}>
+      {({ code, themeName, theme }) => (
+        <Frame
+          code={code}
+          components={components}
+          themeName={themeName}
+          theme={theme}
+          FrameComponent={FrameComponent}
+          ErrorComponent={SendErrorMessage}
+        />
+      )}
+    </UrlParams>,
     outlet
   );
 };
